requirements.txt
click
pyyaml
sh
logbook
staticjinja
markdown
awscli
